Changes needed to correctly represent not analysesd files contained in a SoftwareRelase and their programming language
Changed classes (name, comment and related constraints):
- SoftwareWork: the definition now clarifies that this is common
superclass of both abstract entities (e.g. SoftwareEntity) and concrete
entities (e.g. File) 
- SoftwareRelease: now it is just the set of files released by a
project. The idea that a SoftwareRelease is a SoftwareEntity was already
discarded in the past because a release may contain also files that are
not a manifestation of any SoftwareEntity, moreover if two projects
would release a single SoftwareEntity we would have 2 releases (with two
different release dates), but the SoftwareEntity would be one, so they
cannot be the same. The idea that a SoftwareRelease contains both the
SoftwareEntities and their manifestation would complicate the definition
without clear advantages. The previousVersion of a SoftwareRelease must
be a SoftwareRelease
- ExecutableFile: now is an ObjectCodeFile (to be coherent with the
definition of ObjectCodeFile on turn based on the GPL definition of
ObjectCode). Moreover now the definition is restricted to OS executable
files (the original UML definition was too subjective to be useful for
automatic analyses)
- ConfigurationFile: now is a SourceCodeFile (to be coherent with the
definition of SourceCodeFile on turn based on the GPL definition of
SourceCode)

Added classes (name, comment and related constraints):
- CodeFile: the disjoint union of SourceCodeFile and ObjectCodeFile	(the
most general File type that is coded using a programming language)

Removed classes:
- NonCodeFile: because the definition was inconsistent (as according to
the other definitions a "non SourceCodeFile" is an "ObjectCodeFile" and
a "non (SourceCodeFile o ObjectCodeFile)" is Nothing !

Added properties:
- codingLanguage: CodeFile x ProgrammingLanguage (the programming
language that the code file is written in).
- codingLanguageVersion: CodeFile x ProgrammingLanguageVersion (the
programming language version that the code file is written in).